UC-3 fix: сделал все page серверными
