+++ src/app/auth/password-reset/page.tsx
@@ -1,20 +1,5 @@
-'use client';
+import { PasswordResetClient } from "./components/PasswordResetClient";
 
-import { useSearchParams, redirect } from 'next/navigation';
-import { PATH } from '@/shared/constants/app-paths';
-import { ForgotPasswordConfirmationModal } from '@/shared/ui/modal/components/ForgotPasswordConfirmationModal/ForgotPasswordConfirmationModal';
-import { useState } from 'react';
-
-export default function PasswordResetPage() {
-const [open, setOpen] = useState(true);
-
-  const searchParams = useSearchParams();
-  const email = searchParams.get('email');
-
-  // Если нет email, редиректим на страницу восстановления пароля
-  if (!email) {
-    redirect(PATH.PASSWORD_RECOVERY);
-  }
-
-  return <ForgotPasswordConfirmationModal open={open} onClose={() => setOpen(false)}/>  ;
+export default function NewPasswordResetPage() {
+  return <PasswordResetClient /> 
 }
