+++ src/app/auth/recovery/page.tsx
@@ -1,0 +1,5 @@
+import { RecoveryClient } from './components/RecoveryClient'
+
+export default function NewPasswordPage() {
+  return <RecoveryClient />
+}
